requirements.txt
Pillow==9.1.0
Pillow==9.5.0
